pyproject.toml
[project]
name = "healthcare-research-agent"
version = "0.1.0"
description = "Healthcare research agent using Google ADK A2A"
requires-python = ">=3.13"
dependencies = [
    "agent-framework==1.0.0b251120",
    "agent-framework-a2a==1.0.0b251120",
    "agentstack-sdk==0.4.3",
    "beeai-framework[a2a,vertexai]==0.1.74",
    "google-adk[a2a]==1.19.0",
    "google-generativeai>=0.8.5",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
research-agent = "research_agent.agentstack_reserach:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["research_agent"]
